requirements.txt
# Core Dependencies
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
pymongo==4.6.0

# Environment
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
